Задание №4 - Модуляция (1.0)__Я пока не могу понять что можно еще добавит
